+++ dummy/priv/repo/migrations/20240101000009_add_dashboard_layout.exs
@@ -1,0 +1,6 @@
+defmodule Dummy.Repo.Migrations.AddDashboardLayout do
+  use Ecto.Migration
+
+  def up, do: PhoenixLens.Migrations.up()
+  def down, do: :ok
+end